pulsar-agents/requirements.txt
feedparser==6.0.11
groq==0.11.0
google-cloud-firestore==2.16.1
google-auth==2.29.0
python-dateutil==2.9.0
requests==2.31.0
praw==7.7.1
